# backend/swagger_server/swagger/swagger.yaml
...
swagger: '2.0'
info:
  description: "PDF Service"
  version: "0.0.1"
  title: "PDF Service"
host: "sign.kew.ca"
...
basePath: "/v1"
schemes:
- "https"
paths:
  /document:
    post:
      summary: "Create a document"
      description:
        Verify that all fields that are linked to
        signators actually have corresponding fields
        within the document and that those fields
        are actually fillable (contain a field specifier).
        This method aborts with HTTP 400 if the conditions
        are violated.
      operationId: "document_post"
      consumes:
        - "multipart/form-data"
      parameters:
        - name: "docName"
          in: "formData"
          description: "The name of the document to be created."
          required: true
          type: "string"
        - name: "signators"
          in: "formData"
          description: 
            "JSON object detailing all the signators for the form\
            \ and the fields that they are supposed to be signing."
          type: "string"
          required: true
        - name: "file"
          in: "formData"
          description: "The PDF document to be uploaded"
          required: true
          type: "file"
      responses:
        201:
          description: "A new document was successfully created. The location of the\
            \ new document will be returned in the Location header."
          headers:
            Location:
              type: "string"
            X-Document-Id:
              type: "string"
          schema:
            properties:
              docId:
                type: string
                description: The document ID
        400:
          description: "The document was not a valid PDF form."
          schema:
            $ref: "#/definitions/ErrorMessage"
        401:
          $ref: "#/responses/Unauthorized"
        413:
          $ref: "#/responses/RequestTooLarge"
        415:
          $ref: "#/responses/InvalidContentType"
        422:
          $ref: "#/responses/InvalidToken"
        429:
          $ref: "#/responses/TooManyRequests"
      security:
        - jwt: []
      x-swagger-router-controller: "swagger_server.controllers.document.root"
  /document/{docId}:
    get:
      summary: "View a document"
      description:
        Fetch a document or page of a document.
        This endpoint will record that the document
        has been viewed in the audit log. The
        format that will returned is decided by the
        Accept header. The default format is
        application/pdf in cases where no accept
        header is provided.
      operationId: "get"
      produces:
        - application/pdf
        - image/png
      parameters:
        - name: docId
          in: path
          description: The document ID that the client wishes to view
          required: true
          type: string
        - name: page
          in: query
          description:
            When requesting an image, indicate which page the client
            wishes to request. If the client is requesting a PDF,
            this parameter is ignored.
          required: false
          type: number
      responses:
        200:
          description: "The document has been retrieved successfully"
          schema:
            type: file
        401:
          $ref: "#/responses/Unauthorized"
        404:
          description: "A document with the given ID could not be found."
          schema:
            $ref: "#/definitions/ErrorMessage"
        422:
          $ref: "#/responses/InvalidToken"
        429:
          $ref: "#/responses/TooManyRequests"
      security:
        - jwt: []
      x-swagger-router-controller: "swagger_server.controllers.document.doc_id"
    delete:
      summary: "Delete a document"
      description:
        Delete a document and all artifacts associated
        with the document (audit log, rendered pages,
        file usages, field usages, fields, etc.).
        This action is irreversible and will also
        delete the document without notice for all
        users who have fields to fill within the
        document.
      operationId: "delete"
      parameters:
        - name: "docId"
          in: "path"
          description: "The document ID to delete"
          required: true
          type: "string"
      responses:
        204:
          description: "The document was deleted successfully."
        401:
          $ref: "#/responses/Unauthorized"
        404:
          description: "There is no document with the given ID."
          schema:
            $ref: "#/definitions/ErrorMessage"
        422:
          $ref: "#/responses/InvalidToken"
        429:
          $ref: "#/responses/TooManyRequests"
      security:
        - jwt: []
      x-swagger-router-controller: "swagger_server.controllers.document.doc_id"
...
securityDefinitions:
  jwt:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  ErrorMessage:
    type: object
    properties:
      title: 
        type: string
        description: "Error message describing the error"
      description:
        type: string
        description: "Additional error details"
    required:
      - title
    example:
      title: "Error Message"
      description: "More details..."
...
responses:
  TooManyRequests:
    description: 
      The client has submitted too many requests, 
      and should retry at a later time.
    schema:
      $ref: "#/definitions/ErrorMessage"
  Unauthorized:
    description: 
      The client is not authorized to perform this action. If the client
      has the appropriate credentials to view the resource, they should authenticate
      using those credentials.
    schema:
      $ref: "#/definitions/ErrorMessage"
  InvalidToken:
    description:
      The client passed in an authorization token that
      was not a JWT.
    schema:
      $ref: "#/definitions/ErrorMessage"
  InvalidContentType:
    description:
      The client passed in a request that did not 
      have the required content type.
    schema:
      $ref: "#/definitions/ErrorMessage"
  RequestTooLarge:
    description:
      The request is too large and will not be
      accepted by the server.
    schema:
      $ref: "#/definitions/ErrorMessage"